common logic to indexAccountByPublicKey
